refactor: moved disassembly view Pause/Continue button to debug window
